requirements.txt
djongo==1.3.1
pymongo==3.10.1
numpy>=1.15.1
scipy>=1.1.0
torch>=0.4.0
umap-learn
matplotlib>=2.0.2
webrtcvad
librosa>=0.5.1
tqdm
sounddevice
google-api-python-client
django
django-crispy-forms
django-extensions
kafka-python
pydub
numba==0.48
youtube_dl

